+++ src/Orders/SPR.php
@@ -45,11 +45,16 @@ final class SPR extends UploadOrder
         $signatureData = new UserSignature();
         $this->userSignatureHandler->handle($signatureData, $this->transaction->getDigest());
 
+        $signatureVersion = $this->context->getKeyring()->getUserSignatureAVersion();
+        $dataDigest = $this->orderDataHandler->hash($this->orderData->getContent());
+
         $this->context
             ->setOrderType('SPR')
             ->setTransactionKey($this->transaction->getKey())
             ->setNumSegments($this->transaction->getNumSegments())
-            ->setSignatureData($signatureData);
+            ->setSignatureData($signatureData)
+            ->setSignatureVersion($signatureVersion)
+            ->setDataDigest($dataDigest);
 
         return $this->requestFactory
             ->createRequestBuilderInstance()
@@ -95,7 +100,7 @@ final class SPR extends UploadOrder
                                     );
                             })
                             ->addSignatureData($this->context->getSignatureData(), $this->context->getTransactionKey())
-                            ->addDataDigest($this->context->getKeyring()->getUserSignatureAVersion());
+                            ->addDataDigest($this->context->getSignatureVersion(), $this->context->getDataDigest());
                     });
                 });
             })